+++ src/index-module.ts
@@ -1,6 +1,4 @@
 export * from './kit'
-export * from './kit.types'
 export * from './plugins'
 export * from './session'
-export * from './session.types'
 export * from './types'
