+++ src/plugins/login.ts
@@ -1,0 +1,7 @@
+import {AbstractLoginPlugin} from '../types'
+
+export class BaseLoginPlugin extends AbstractLoginPlugin {
+    register() {
+        // console.log('Register hooks via context.addHook')
+    }
+}
