+++ src/plugins/transact.ts
@@ -1,0 +1,7 @@
+import {AbstractTransactPlugin} from '../types'
+
+export class BaseTransactPlugin extends AbstractTransactPlugin {
+    register() {
+        // console.log('Register hooks via context.addHook')
+    }
+}
