+++ src/types.ts
@@ -1,2 +1,15 @@
-export interface Hook {}
+import {Checksum256, Checksum256Type, Struct} from '@greymass/eosio'
+
 export type Fetch = (input: any, init?: any) => Promise<any>
+
+@Struct.type('chain_definition')
+export class ChainDefinition extends Struct {
+    @Struct.field('checksum256') declare id: Checksum256
+    @Struct.field('string') declare url: string
+}
+
+export type ChainDefinitionType = ChainDefinition | {id: Checksum256Type; url: string}
+
+export interface Hook {}
+
+export * from './types/index'
