+++ src/types/hooks/index.ts
@@ -1,3 +1,2 @@
 export * from './login'
 export * from './transact'
-export * from './wallet/privatekey'
