+++ src/types/plugins/transact.ts
@@ -1,0 +1,9 @@
+import {TransactContext} from '../contexts'
+
+export interface TransactPlugin {
+    register: (context: TransactContext) => void
+}
+
+export abstract class AbstractTransactPlugin implements TransactPlugin {
+    public abstract register(context: TransactContext): void
+}
