+++ src/types/transact.ts
@@ -1,0 +1,73 @@
+import {
+    AnyAction,
+    AnyTransaction,
+    Checksum256Type,
+    PermissionLevel,
+    Signature,
+} from '@greymass/eosio'
+
+import {ResolvedSigningRequest, ResolvedTransaction, SigningRequest} from 'eosio-signing-request'
+
+import {ChainDefinition} from '../types'
+import {AbstractTransactPlugin} from './plugins'
+
+/**
+ * Payload accepted by the [[Session.transact]] method.
+ * Note that one of `action`, `actions` or `transaction` must be set.
+ */
+export interface TransactArgs {
+    /** Full transaction to sign. */
+    transaction?: AnyTransaction
+    /** Action to sign. */
+    action?: AnyAction
+    /** Actions to sign. */
+    actions?: AnyAction[]
+    /** An ESR payload */
+    request?: SigningRequest | string
+}
+
+/**
+ * Options for the [[Session.transact]] method.
+ */
+export interface TransactOptions {
+    /**
+     * Whether to broadcast the transaction or just return the signature.
+     * Defaults to true.
+     */
+    broadcast?: boolean
+    /**
+     * Chain to use when configured with multiple chains.
+     */
+    chain?: Checksum256Type
+    /**
+     * Whether to allow the signer to make modifications to the request
+     * (e.g. applying a cosigner action to pay for resources).
+     *
+     * Defaults to true if [[broadcast]] is true or unspecified; otherwise false.
+     */
+    allowModify?: boolean
+    /**
+     * Specific transact plugins to use for this transaction.
+     */
+    transactPlugins?: AbstractTransactPlugin[]
+}
+
+/**
+ * The response from a [[Session.transact]] call.
+ */
+export interface TransactResult {
+    /** The chain that was used. */
+    chain: ChainDefinition
+    /** The SigningRequest representation of the transaction. */
+    request: SigningRequest
+    /** The ResolvedSigningRequest of the transaction */
+    resolved: ResolvedSigningRequest | undefined
+    /** The transaction signatures. */
+    signatures: Signature[]
+    /** The signer authority. */
+    signer: PermissionLevel
+    /** The resulting transaction. */
+    transaction: ResolvedTransaction | undefined
+    /** Push transaction response from api node, only present if transaction was broadcast. */
+    processed?: {[key: string]: any}
+}
